requirements.txt
-e ../shared/python
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
httpx==0.27.2
langgraph==0.2.39
google-generativeai==0.8.5
